+++ src/com/yandex/app/service/FileBackedTaskManager.java
@@ -11,181 +11,23 @@ import java.nio.file.Files;
 import java.time.Duration;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
-import java.util.ArrayList;
-import java.util.Comparator;
 import java.util.List;
-import java.util.TreeSet;
 
 public class FileBackedTaskManager extends InMemoryTaskManager {
     private final File file;
     private HistoryManager history;
-    private TreeSet<Task> prioritizedTasks;
 
     public FileBackedTaskManager(File file, HistoryManager history) {
         super(history);
         this.file = file;
-        prioritizedTasks = new TreeSet<>(
-                Comparator.comparing(Task::getStartTime, Comparator.nullsLast(Comparator.naturalOrder()))
-                        .thenComparing(Task::getId)
-        );
-    }
-
-    @Override
-    public int createTask(Task task) {
-        boolean hasCrossTasks = getAllTasks().stream()
-                .anyMatch(existingTask -> isCrossTasks(existingTask, task));
-        if (hasCrossTasks) {
-            throw new IllegalArgumentException("Есть пересечение с другой задачей");
-        }
-        int id = super.createTask(task);
-        if (task.getStartTime() != null) {
-            prioritizedTasks.add(task);
-        }
-        save();
-        return id;
-    }
-
-    @Override
-    public int createEpic(Epic epic) {
-        int id = super.createEpic(epic);
-        save();
-        return id;
-    }
-
-    @Override
-    public int createSubtask(int epicId, Subtask subtask) {
-        boolean hasCrossTasks = getAllSubtasks().stream()
-                .anyMatch(existingSubtask -> isCrossTasks(existingSubtask, subtask));
-        if (hasCrossTasks) {
-            throw new IllegalArgumentException("Есть пересечение с другой подзадачей");
-        }
-        int id = super.createSubtask(epicId, subtask);
-        if (subtask.getStartTime() != null) {
-            prioritizedTasks.add(subtask);
-        }
-        save();
-        return id;
-    }
-
-    @Override
-    public void updateTask(int id, Task task) {
-        Task oldTask = getTasks().get(id);
-        if (oldTask != null) {
-            prioritizedTasks.remove(oldTask);
-
-            boolean hasCrossTasks = getAllTasks().stream()
-                    .filter(existingTask -> existingTask.getId() != id)
-                    .anyMatch(existingTask -> isCrossTasks(existingTask, task));
-            if (hasCrossTasks) {
-                throw new IllegalArgumentException("Обновлённая задача пересекается с другой задачей");
-            }
-
-            if (task.getStartTime() != null) {
-                prioritizedTasks.add(task);
-            }
-        }
-        super.updateTask(id, task);
-        save();
-    }
-
-    @Override
-    public void updateEpic(int id, Epic epic) {
-        super.updateEpic(id, epic);
-        save();
-    }
-
-    @Override
-    public void updateSubtask(int id, Subtask subtask) {
-        Subtask oldSubtask = getAllSubtasks().get(id);
-        if (oldSubtask != null) {
-            prioritizedTasks.remove(oldSubtask);
-
-            if (hasCrossingTasks(id, subtask)) {
-                throw new IllegalArgumentException("Обновлённая подзадача пересекается с другой подзадачей");
-            }
-
-            if (subtask.getStartTime() != null) {
-                prioritizedTasks.add(subtask);
-            }
-        }
-        super.updateSubtask(id, subtask);
-        save();
-    }
-
-    private boolean hasCrossingTasks(int id, Subtask subtask) {
-        return getAllSubtasks().stream()
-                .filter(existingSubtask -> existingSubtask.getId() != id)
-                .anyMatch(existingSubtask -> isCrossTasks(existingSubtask, subtask));
-    }
-
-    @Override
-    public boolean deleteTaskById(int id) {
-        Task task = getTaskById(id);
-        boolean result = super.deleteTaskById(id);
-        if (result && task != null) {
-            prioritizedTasks.remove(task);
-            save();
-        }
-        return result;
-    }
-
-    @Override
-    public boolean deleteSubtaskById(int id) {
-        Subtask subtask = getSubtaskById(id);
-        boolean result = super.deleteSubtaskById(id);
-        if (result && subtask != null) {
-            prioritizedTasks.remove(subtask);
-            save();
-        }
-        return result;
-    }
-
-    @Override
-    public boolean deleteEpicById(int id) {
-        Epic epic = getEpicById(id);
-        boolean result = super.deleteEpicById(id);
-        if (result) {
-            if (getAllSubtasksOfEpic(epic.getId()) != null) {
-                for (Subtask subtask : getAllSubtasksOfEpic(epic.getId())) {
-                    prioritizedTasks.remove(subtask);
-                }
-            }
-            if (epic != null) {
-                prioritizedTasks.remove(epic);
-            }
-            save();
-        }
-        return result;
-    }
-
-    public boolean isCrossTasks(Task task1, Task task2) {
-        LocalDateTime start1 = task1.getStartTime();
-        LocalDateTime end1 = task1.getEndTime();
-        LocalDateTime start2 = task2.getStartTime();
-        LocalDateTime end2 = task2.getEndTime();
-
-        return start1.isBefore(end2) && start2.isBefore(end1);
-    }
-
-    public List<Task> getPrioritizedTasks() {
-        return new ArrayList<>(prioritizedTasks);
-    }
-
-    public void checkCrossTasks() {
-        List<Task> tasks = getPrioritizedTasks();
-        for (int i = 0; i < tasks.size() - 1; i++) {
-            Task currentTask = tasks.get(i);
-            Task nextTask = tasks.get(i + 1);
-
-            if (currentTask.getEndTime().isAfter(nextTask.getStartTime())) {
-                System.out.println("Задачи " + currentTask + " и " + nextTask + " пересекаются");
-            }
-        }
     }
 
     public void save() {
         File file = new File("tasks.csv");
         try (BufferedWriter writer = new BufferedWriter(new FileWriter(file))) {
+            writer.write("Id,Type,Title,Status,Description,Duration,StartTime,EndTime");
+            writer.newLine();
+
             for (Task task : getAllTasks()) {
                 writer.write(task.toString());
                 writer.newLine();
@@ -247,8 +89,14 @@ public class FileBackedTaskManager extends InMemoryTaskManager {
             System.out.println("Absolute path: " + file.getAbsolutePath());
             String strings = Files.readString(file.toPath(), StandardCharsets.UTF_8);
             String[] lines = strings.split(System.lineSeparator());
+
+            boolean isFirstLine = true;
             for (String line : lines) {
                 if (line.trim().isEmpty()) {
+                    continue;
+                }
+                if (isFirstLine) {
+                    isFirstLine = false; // Пропускаем первую строку (заголовок)
                     continue;
                 }
                 if (line.startsWith("HISTORY:")) {
