+++ src/com/yandex/app/service/InMemoryTaskManager.java
@@ -4,22 +4,33 @@ import com.yandex.app.model.Epic;
 import com.yandex.app.model.Subtask;
 import com.yandex.app.model.Task;
 
+import java.time.LocalDateTime;
 import java.util.*;
 
 public class InMemoryTaskManager implements TaskManager {
     private Map<Integer, Task> tasks;
     private Map<Integer, Subtask> subtasks;
     private Map<Integer, Epic> epics;
+    private HistoryManager history;
+    protected TreeSet<Task> prioritizedTasks;
+
     private int currentTaskId = 0;
     private int currentSubtaskId = 100;
     private int currentEpicId = 1000;
-    private HistoryManager history;
 
     public InMemoryTaskManager(HistoryManager history) {
         tasks = new HashMap<>();
         subtasks = new HashMap<>();
         epics = new HashMap<>();
         this.history = history;
+        prioritizedTasks = new TreeSet<>(
+                Comparator.comparing(Task::getStartTime, Comparator.nullsLast(Comparator.naturalOrder()))
+                        .thenComparing(Task::getId)
+        );
+    }
+
+    public List<Task> getPrioritizedTasks() {
+        return new ArrayList<>(prioritizedTasks);
     }
 
     // методы для задач
@@ -50,9 +61,16 @@ public class InMemoryTaskManager implements TaskManager {
 
     @Override
     public int createTask(Task task) {
+        if (hasCrossingTasks(task)) {
+            throw new IllegalArgumentException("Есть пересечение с другой задачей");
+        }
         int id = ++currentTaskId;
         tasks.put(id, task);
         history.updateHistory(task);
+        if (task.getStartTime() != null) {
+            prioritizedTasks.add(task);
+        }
+        save();
         return id;
     }
 
@@ -67,25 +85,31 @@ public class InMemoryTaskManager implements TaskManager {
 
     @Override
     public void updateTask(int id, Task task) {
-        if (tasks.containsKey(id)) {
-            tasks.replace(id, task);
+        Task oldTask = getTasks().get(id);
+        if (oldTask == null) {
+            throw new IllegalArgumentException("Задача с указанным id не существует");
         }
+        prioritizedTasks.remove(oldTask);
+        if (hasCrossingTasks(task)) {
+            throw new IllegalArgumentException("Обновлённая задача пересекается с другой задачей");
+        }
+        if (task.getStartTime() != null) {
+            prioritizedTasks.add(task);
+        }
+        tasks.put(id, task);
+        save();
     }
 
     @Override
     public boolean deleteTaskById(int id) {
-        Task task = tasks.get(id);
-        if (task == null) {
-            return false;
-        } else {
-            boolean isRemoved = tasks.remove(id) != null;
+        Task task = tasks.remove(id);
+        if (task != null) {
+            prioritizedTasks.remove(task);
             history.remove(id);
-            if (isRemoved) {
-                return true;
-            } else {
-                return false;
-            }
+            save();
+            return true;
         }
+        return false;
     }
 
     @Override
@@ -110,6 +134,9 @@ public class InMemoryTaskManager implements TaskManager {
 
     @Override
     public int createSubtask(int epicId, Subtask subtask) {
+        if (hasCrossingTasks(subtask)) {
+            throw new IllegalArgumentException("Есть пересечение с другой подзадачей");
+        }
         int id = ++currentSubtaskId;
         subtasks.put(id, subtask);
         Epic epic = getEpicById(epicId);
@@ -119,6 +146,10 @@ public class InMemoryTaskManager implements TaskManager {
             epic.calculateEpicDuration();
             history.updateHistory(subtask);
         }
+        if (subtask.getStartTime() != null) {
+            prioritizedTasks.add(subtask);
+        }
+        save();
         return id;
     }
 
@@ -136,39 +167,45 @@ public class InMemoryTaskManager implements TaskManager {
 
     @Override
     public void updateSubtask(int id, Subtask newSubtask) {
-        if (subtasks.containsKey(id)) {
-            Subtask oldSubtask = subtasks.get(id);
-            subtasks.replace(id, newSubtask);
-
-            Integer epicId = oldSubtask.getEpicId();
-            if (epicId != null) {
-                Epic epic = getEpicById(id);
-                if (epic != null) {
-                    epic.updateEpicStatus();
-                }
+        Subtask oldSubtask = getAllSubtasks().get(id);
+        if (oldSubtask == null) {
+            throw new IllegalArgumentException("Подзадача с указанным id не существует");
+        }
+        prioritizedTasks.remove(oldSubtask);
+        if (hasCrossingTasks(newSubtask)) {
+            throw new IllegalArgumentException("Обновлённая задача пересекается с другой подзадачей");
+        }
+        if (newSubtask.getStartTime() != null) {
+            prioritizedTasks.add(newSubtask);
+        }
+        subtasks.put(id, newSubtask);
+        Integer epicId = newSubtask.getEpicId();
+        if (epicId != null) {
+            Epic epic = getEpicById(id);
+            if (epic != null) {
+                epic.updateEpicStatus();
             }
         }
+        save();
     }
 
     @Override
     public boolean deleteSubtaskById(int id) {
-        Subtask subtask = subtasks.get(id);
+        Subtask subtask = getSubtaskById(id);
         if (subtask == null) {
             return false;
-        } else {
-            boolean isRemoved = subtasks.remove(id) != null;
-            history.remove(id);
-            Epic epic = getEpicById(subtask.getEpicId());
-            if (epic != null) {
-                epic.getSubtaskIds().removeIf(taskId -> taskId == id);
-                epic.updateEpicStatus();
-            }
-            if (isRemoved) {
-                return true;
-            } else {
-                return false;
-            }
         }
+        boolean isRemoved = subtasks.remove(id) != null;
+        history.remove(id);
+        prioritizedTasks.remove(subtask);
+
+        Epic epic = getEpicById(subtask.getEpicId());
+        if (epic != null) {
+            epic.getSubtaskIds().removeIf(taskId -> taskId == id);
+            epic.updateEpicStatus();
+        }
+        save();
+        return isRemoved;
     }
 
     @Override
@@ -201,6 +238,7 @@ public class InMemoryTaskManager implements TaskManager {
         int id = ++currentEpicId;
         epics.put(id, epic);
         history.updateHistory(epic);
+        save();
         return id;
     }
 
@@ -221,6 +259,7 @@ public class InMemoryTaskManager implements TaskManager {
         if (epics.containsKey(id)) {
             epics.replace(id, epic);
         }
+        save();
     }
 
     @Override
@@ -234,15 +273,52 @@ public class InMemoryTaskManager implements TaskManager {
         subtasksOfEpic.forEach(subtask -> {
             deleteSubtaskFromEpic(subtask.getId());
             history.remove(subtask.getId());
+            prioritizedTasks.remove(subtask);
         });
 
+        prioritizedTasks.remove(epic);
         boolean isRemoved = epics.remove(id) != null;
         history.remove(id);
+
+        save();
         return isRemoved;
     }
 
     @Override
     public void deleteAllEpics() {
         epics.keySet().forEach(this::deleteEpicById);
+    }
+
+    protected boolean hasCrossingTasks(Task task) {
+        return prioritizedTasks.stream()
+                .anyMatch(existingTask -> isCrossTasks(existingTask, task));
+    }
+
+    public boolean isCrossTasks(Task task1, Task task2) {
+        LocalDateTime start1 = task1.getStartTime();
+        LocalDateTime end1 = task1.getEndTime();
+        LocalDateTime start2 = task2.getStartTime();
+        LocalDateTime end2 = task2.getEndTime();
+
+        if (start1 == null || end1 == null || start2 == null || end2 == null) {
+            return false;
+        }
+
+        return start1.isBefore(end2) && start2.isBefore(end1);
+    }
+
+    public void checkCrossTasks() {
+        List<Task> tasks = getPrioritizedTasks();
+        for (int i = 0; i < tasks.size() - 1; i++) {
+            Task currentTask = tasks.get(i);
+            Task nextTask = tasks.get(i + 1);
+
+            if (currentTask.getEndTime().isAfter(nextTask.getStartTime())) {
+                System.out.println("Задачи " + currentTask + " и " + nextTask + " пересекаются");
+            }
+        }
+    }
+
+    protected void save() {
     }
 }
